Use grouped test name options to form
